kafka: move constants and print helpers into protocol
These types and the print helpers are part of the protocol.
